3.19

+++ src/main/java/com/edu/ctbu/demo/domain/Members.java
@@ -20,7 +20,7 @@ public class Members {
 
     private String membername;
 
-    private Long gender;
+    private Integer gender;
 
     private Float originalamount;//原始金额
